Add GhostCat CVE 2020-1938
