+++ src/app/sitemap.php
@@ -32,3 +32,15 @@ $app->get(
       return $response->withHeader("content-type", "application/xml");
   }
 );
+
+$app->get(
+  '/',
+  function (Request $request, Response $response, array $args) {
+      $response->getBody()->write(
+        "Welcome to HypeJabdsadsa! 💉 😃 <br>"
+        . "HypeJab is a deliberately vulnerable web application"
+        . " intended for benchmarking automated scanners."
+      );
+      return $response;
+  }
+);
