+++ src/app/vulnerabilities/GhostCatCVE20201938.php
@@ -1,0 +1,15 @@
+<?php
+
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
+use Slim\Factory\AppFactory;
+
+$app->get(
+    '/conf/server.xml',
+    function (Request $request, Response $response) {
+        $xml = file_get_contents(__DIR__ . "/../resources/server.xml");
+        $response->getBody()->write($xml);
+        return $response->withHeader("content-type", "application/xml")
+                        ->withStatus(200);
+    }
+  );
